[examples] Extract common Linux platform app start up code (#6222)
* [examples] Extract common Linux platform app start up code

* Restyled by gn

* Use LogDeviceConfig + PrintOnboardingCodes

* Move PW_RPC_ENABLED back to program

* Add PW_RPC_ENABLED

* Fix Linux build
