+++ examples/lighting-app/linux/BUILD.gn
@@ -31,13 +31,13 @@ config("includes") {
 executable("chip-lighting-app") {
   sources = [
     "LightingManager.cpp",
-    "Options.cpp",
     "include/LightingManager.h",
     "main.cpp",
   ]
 
   deps = [
     "${chip_root}/examples/lighting-app/lighting-common",
+    "${chip_root}/examples/platform/linux:app-main",
     "${chip_root}/src/lib",
   ]
