+++ examples/platform/linux/BUILD.gn
@@ -12,23 +12,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import("//build_overrides/build.gni")
 import("//build_overrides/chip.gni")
+import("${chip_root}/examples/common/pigweed/pigweed_rpcs.gni")
 
-executable("chip-all-clusters-app") {
-  sources = [ "main.cpp" ]
-
-  deps = [
-    "${chip_root}/examples/all-clusters-app/all-clusters-common",
-    "${chip_root}/examples/platform/linux:app-main",
-    "${chip_root}/src/lib",
-  ]
-
-  cflags = [ "-Wconversion" ]
-
-  output_dir = root_out_dir
+config("app-main-config") {
+  include_dirs = [ "." ]
 }
 
-group("linux") {
-  deps = [ ":chip-all-clusters-app" ]
+source_set("app-main") {
+  sources = [
+    "AppMain.cpp",
+    "AppMain.h",
+    "Options.cpp",
+    "Options.h",
+  ]
+
+  if (chip_enable_pw_rpc) {
+    defines = [ "PW_RPC_ENABLED" ]
+  }
+
+  public_deps = [ "${chip_root}/src/lib" ]
+
+  public_configs = [ ":app-main-config" ]
 }
