+++ examples/tv-app/linux/BUILD.gn
@@ -30,6 +30,7 @@ executable("chip-tv-app") {
   ]
 
   deps = [
+    "${chip_root}/examples/platform/linux:app-main",
     "${chip_root}/examples/tv-app/tv-common",
     "${chip_root}/src/lib",
     "${chip_root}/third_party/inipp",
